Refactor4 Encoding.java Extract_Class
