+++ src/main/java/io/nats/client/support/Base32Utils.java
@@ -1,0 +1,88 @@
+package io.nats.client.support;
+
+import java.util.Arrays;
+
+public class Base32Utils {
+    // http://en.wikipedia.org/wiki/Base_32
+    private static final String BASE32_CHARS = "ABCDEFGHIJKLMNOPQRSTUVWXYZ234567";
+    private static final int[] BASE32_LOOKUP;
+    private static final int MASK = 31;
+    private static final int SHIFT = 5;
+    public static char[] base32Encode(final byte[] input) {
+        int last = input.length;
+        char[] charBuff = new char[(last + 7) * 8 / SHIFT];
+        int offset = 0;
+        int buffer = input[offset++];
+        int bitsLeft = 8;
+        int i = 0;
+
+        while (bitsLeft > 0 || offset < last) {
+            if (bitsLeft < SHIFT) {
+                if (offset < last) {
+                    buffer <<= 8;
+                    buffer |= (input[offset++] & 0xff);
+                    bitsLeft += 8;
+                } else {
+                    int pad = SHIFT - bitsLeft;
+                    buffer <<= pad;
+                    bitsLeft += pad;
+                }
+            }
+            int index = MASK & (buffer >> (bitsLeft - SHIFT));
+            bitsLeft -= SHIFT;
+            charBuff[i] = BASE32_CHARS.charAt(index);
+            i++;
+        }
+
+        int nonBlank;
+
+        for (nonBlank=charBuff.length-1;nonBlank>=0;nonBlank--) {
+            if (charBuff[nonBlank] != 0) {
+                break;
+            }
+        }
+
+        char[] retVal = new char[nonBlank+1];
+
+        System.arraycopy(charBuff, 0, retVal, 0, retVal.length);
+
+        Arrays.fill(charBuff, '\0');
+
+        return retVal;
+    }
+    static {
+        BASE32_LOOKUP = new int[256];
+
+        Arrays.fill(BASE32_LOOKUP, 0xFF);
+
+        for (int i = 0; i < BASE32_CHARS.length(); i++) {
+            int index = BASE32_CHARS.charAt(i) - '0';
+            BASE32_LOOKUP[index] = i;
+        }
+    }
+
+    public static byte[] base32Decode(final char[] input) {
+        byte[] bytes = new byte[input.length * SHIFT / 8];
+        int buffer = 0;
+        int next = 0;
+        int bitsLeft = 0;
+
+        for (char value : input) {
+            int lookup = value - '0';
+
+            if (lookup < 0 || lookup >= BASE32_LOOKUP.length) {
+                continue;
+            }
+
+            int c = BASE32_LOOKUP[lookup];
+            buffer <<= SHIFT;
+            buffer |= c & MASK;
+            bitsLeft += SHIFT;
+            if (bitsLeft >= 8) {
+                bytes[next++] = (byte) (buffer >> (bitsLeft - 8));
+                bitsLeft -= 8;
+            }
+        }
+        return bytes;
+    }
+}
